split outbound bridge impl
